X2EngineCRM 2.1 Changelog 10/12/2012 - Overhauled real-time notification and chat 	- Much lower server load, especially with multiple tabs - Improved URL handling (more efficient) - Improved changelog storage - Big improvements to the installer 	- Real-time installation status updates 	- No more timeout errors - Numerous bug fixes - Improvements to Relationships for contacts, accounts and opportunities - Fix for all bugs related to browsers caching old javascript files - Additional feature in Customization Framework: you can now override controller   files by adding "My" to the class and putting the file in /custom, for example   to override actionIndex in ContactsController you can create a class   MyContactsController extending ContactsController and only define actionIndex.   This class will automatically be used in place of the original file, and you   don't have to override the entire class.

+++ x2engine/protected/controllers/ProfileController.php
@@ -247,7 +247,7 @@ class ProfileController extends x2base {
 			$prof=ProfileChild::model()->findByPk($id);
 			if(isset($_FILES['photo'])){
 				if ($_FILES["photo"]["size"] < 2000000){
-					if($prof->avatar!='' && isset($prof->avatar)){
+					if($prof->avatar!='' && isset($prof->avatar) && file_exists($prof->avatar)){
 						unlink($prof->avatar);
 					}
 					$temp = CUploadedFile::getInstanceByName('photo');
